# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriSmart – Intelligent Farm Management and Decision Support System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farms:
    get:
      operationId: listFarms
      tags: [farms]
      summary: List all farms for the current user
      responses:
        "200":
          description: List of farms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    post:
      operationId: createFarm
      tags: [farms]
      summary: Create a new farm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmInput"
      responses:
        "201":
          description: Created farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{id}:
    get:
      operationId: getFarm
      tags: [farms]
      summary: Get a farm by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Farm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "404":
          description: Not found
    put:
      operationId: updateFarm
      tags: [farms]
      summary: Update a farm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmInput"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
        "404":
          description: Not found
    delete:
      operationId: deleteFarm
      tags: [farms]
      summary: Delete a farm
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Crops ───────────────────────────────────────────────────────────────────
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses with optional filters
      parameters:
        - name: farmId
          in: query
          schema:
            type: number
        - name: cropId
          in: query
          schema:
            type: number
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/summary:
    get:
      operationId: getExpenseSummary
      tags: [expenses]
      summary: Get monthly and yearly expense summaries
      parameters:
        - name: farmId
          in: query
          schema:
            type: number
        - name: year
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Expense summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSummary"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get an expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Expense details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
    put:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Updated expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # ─── Disease Detection ───────────────────────────────────────────────────────
...
  /fertilizer-recommendations:
    get:
      operationId: listFertilizerRecommendations
      tags: [fertilizer]
      summary: List fertilizer recommendation history
      parameters:
        - name: farmId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Recommendation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FertilizerRecommendation"
    post:
      operationId: createFertilizerRecommendation
      tags: [fertilizer]
      summary: Get fertilizer recommendation based on soil data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FertilizerRecommendationInput"
      responses:
        "201":
          description: Recommendation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FertilizerRecommendation"
  /fertilizer-recommendations/{id}:
    get:
      operationId: getFertilizerRecommendation
      tags: [fertilizer]
      summary: Get a fertilizer recommendation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Recommendation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FertilizerRecommendation"
        "404":
          description: Not found

  # ─── Market Prices ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Farm:
      type: object
      required: [id, userId, name, location, area, soilType, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        name:
          type: string
        location:
          type: string
        area:
          type: number
          description: Area in acres
        soilType:
          type: string
          enum: [clay, sandy, loamy, silty, peaty, chalky]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        cropsCount:
          type: number
    FarmInput:
      type: object
      required: [name, location, area, soilType]
      properties:
        name:
          type: string
          minLength: 1
        location:
          type: string
          minLength: 1
        area:
          type: number
          minimum: 0
        soilType:
          type: string
          enum: [clay, sandy, loamy, silty, peaty, chalky]
        description:
          type: string

    # ── Crops ──────────────────────────────────────────────────────────────────
...
    Expense:
      type: object
      required: [id, farmId, category, amount, date, createdAt]
      properties:
        id:
          type: number
        farmId:
          type: number
        cropId:
          type: ["number", "null"]
        category:
          type: string
          enum: [seeds, fertilizers, pesticides, labor, irrigation, equipment, other]
        amount:
          type: number
        description:
          type: ["string", "null"]
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [farmId, category, amount, date]
      properties:
        farmId:
          type: number
        cropId:
          type: number
        category:
          type: string
          enum: [seeds, fertilizers, pesticides, labor, irrigation, equipment, other]
        amount:
          type: number
          minimum: 0
        description:
          type: string
        date:
          type: string
          format: date
    ExpenseSummary:
      type: object
      required: [totalYear, byMonth, byCategory]
      properties:
        totalYear:
          type: number
        byMonth:
          type: array
          items:
            type: object
            required: [month, total]
            properties:
              month:
                type: string
              total:
                type: number
        byCategory:
          type: array
          items:
            type: object
            required: [category, total]
            properties:
              category:
                type: string
              total:
                type: number

    # ── Disease Detection ────────────────────────────────────────────────────
...
    FertilizerRecommendation:
      type: object
      required: [id, farmId, soilType, nitrogen, phosphorus, potassium, ph, fertilizerName, dosage, schedule, createdAt]
      properties:
        id:
          type: number
        farmId:
          type: number
        soilType:
          type: string
        nitrogen:
          type: number
        phosphorus:
          type: number
        potassium:
          type: number
        ph:
          type: number
        cropName:
          type: ["string", "null"]
        fertilizerName:
          type: string
        dosage:
          type: string
        schedule:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    FertilizerRecommendationInput:
      type: object
      required: [farmId, soilType, nitrogen, phosphorus, potassium, ph]
      properties:
        farmId:
          type: number
        soilType:
          type: string
          enum: [clay, sandy, loamy, silty, peaty, chalky]
        nitrogen:
          type: number
          minimum: 0
          maximum: 140
        phosphorus:
          type: number
          minimum: 5
          maximum: 145
        potassium:
          type: number
          minimum: 5
          maximum: 205
        ph:
          type: number
          minimum: 0
          maximum: 14
        cropName:
          type: string

    # ── Market Prices ──────────────────────────────────────────────────────
...