# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AgriSmart – Intelligent Farm Management and Decision Support System API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseCategory } from './expenseCategory';

export interface Expense {
  id: number;
  farmId: number;
  /** @nullable */
  cropId?: number | null;
  category: ExpenseCategory;
  amount: number;
  /** @nullable */
  description?: string | null;
  date: Date;
  createdAt: Date;
}