# lib/api-zod/src/generated/types/fertilizerRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AgriSmart – Intelligent Farm Management and Decision Support System API
 * OpenAPI spec version: 0.1.0
 */

export interface FertilizerRecommendation {
  id: number;
  farmId: number;
  soilType: string;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  ph: number;
  /** @nullable */
  cropName?: string | null;
  fertilizerName: string;
  dosage: string;
  schedule: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}